requirements.txt
acme==4.2.0
amqp==5.3.1
asgiref==3.8.1
asymmetric-jwt-auth==1.2.0
billiard==4.2.1
celery==5.5.3
certbot==4.2.0
certbot-django==0.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
ConfigArgParse==1.7.1
configobj==5.0.9
cron-descriptor==1.4.5
cryptography==45.0.3
diff-match-patch==20241021
distro==1.9.0
Django==4.2.23
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-import-export==4.3.8
django-json-widget==2.0.1
django-jsonform==2.23.2
django-mongodb-engine==0.6.0
django-ranged-response==0.2.0
django-redis==6.0.0
django-simple-captcha==0.6.2
django-timezone-field==7.1
django_neomodel==0.2.0
django_tiptap==0.0.18
djangorestframework==3.16.0
djangotoolbox==1.8.0
dnspython==2.7.0
docker==7.1.0
gunicorn==23.0.0
idna==3.10
joblib==1.5.1
josepy==2.1.0
kombu==5.5.4
Markdown==3.8
mongoengine==0.29.1
neo4j==5.19.0
neomodel==5.3.3
numpy==2.3.1
packaging==25.0
parsedatetime==2.6
pillow==11.2.1
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.10.1
pymongo==4.14.0
pyOpenSSL==25.1.0
pyRFC3339==2.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
redis==6.2.0
requests==2.32.3
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
